move path-directory-making to path package
